several improvements to make the code more readable

+++ TestPrograms/fs-test/src/Flash/flashHAL.cpp
@@ -3,12 +3,12 @@
 uint8_t* FlashHAL::flash_memory_pointer = nullptr;
 
 void FlashHAL::flash_range_erase(uint32_t flash_offs, size_t count){
-    if(calculateSectorAddress(calculateSector((int)flash_offs)) != (int)flash_offs) {
-      printf("ERROR: Flash erase start address is not aligned to sector start: %u\n", (int)flash_offs);
+    if(calculateSectorAddress(calculateSector(flash_offs)) != flash_offs) {
+      printf("ERROR: Flash erase start address is not aligned to sector start: %i\n", flash_offs);
       return; // not aligned to sector start
     }
-    if(calculateSectorAddress(calculateSector((int)count)) != (int)count) {
-      printf("ERROR: Flash erase size is not aligned to sector size: %u\n", (int)count);
+    if(calculateSectorAddress(calculateSector(count)) != count) {
+      printf("ERROR: Flash erase size is not aligned to sector size: %li\n", count);
       return; // not aligned to sector size
     }
     std::memset((uint8_t*)getFlashMemoryOffset() + flash_offs, 0xFF, count);
@@ -16,12 +16,12 @@ void FlashHAL::flash_range_erase(uint32_t flash_offs, size_t count){
 
 void FlashHAL::flash_range_program(uint32_t flash_offs, const uint8_t *data,
                                   size_t count){
-    if(calculatePageAddress(calculatePage((int)flash_offs)) != (int)flash_offs) {
-      printf("ERROR: Flash program start address is not aligned to page start: %u\n", (int)flash_offs);
+    if(calculatePageAddress(calculatePage(flash_offs)) != flash_offs) {
+      printf("ERROR: Flash program start address is not aligned to page start: %i\n", flash_offs);
       return; // not aligned to page start
     }
-    if(calculatePageAddress(calculatePage((int)count)) != (int)count) {
-      printf("ERROR: Flash program size is not aligned to page size: %u\n", (int)count);
+    if(calculatePageAddress(calculatePage(count)) != count) {
+      printf("ERROR: Flash program size is not aligned to page size: %li\n", count);
       return; // not aligned to page size
     }
